# gen/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ads app API Documentation
  version: "1.0"
servers:
- url: http://localhost:8080
paths:
...
  /register:
    post:
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterReq'
      responses:
        "404":
          description: Not Found
        "201":
          description: Created
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      summary: Регистрация пользователя
      tags:
      - Регистрация
      x-contentType: application/json
      x-accepts: application/json
  /login:
    post:
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginReq'
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      summary: Авторизация пользователя
      tags:
      - Авторизация
      x-contentType: application/json
      x-accepts: application/json
...
components:
...
  schemas:
...
    RegisterReq:
      example:
        firstName: firstName
        lastName: lastName
        password: password
        role: USER
        phone: phone
        username: username
      properties:
        username:
          description: логин
          type: string
        password:
          description: пароль
          type: string
        firstName:
          description: имя пользователя
          type: string
        lastName:
          description: фамилия пользователя
          type: string
        phone:
          description: телефон пользователя
          type: string
        role:
          description: роль пользователя
          enum:
          - USER
          - ADMIN
          type: string
      type: object
    LoginReq:
      example:
        password: password
        username: username
      properties:
        password:
          description: пароль
          type: string
        username:
          description: логин
          type: string
      type: object
...