# src/main/java/ru/skypro/project/marketpkace/dto/RegisterReq.java
package ru.skypro.project.marketpkace.dto;

import lombok.Data;

@Data
public class RegisterReq {
    private String username;
    private String password;
    private String firstName;
    private String lastName;
    private String phone;
    private Role role;
}